+++ server/src/modules/users/user.controller.ts
@@ -7,15 +7,7 @@ class UserController {
   constructor(
     private userService: UserService,
     private fileService: FileService
-  ) {
-    this.getAllUsers = this.getAllUsers.bind(this);
-    this.getUserById = this.getUserById.bind(this);
-    this.createUser = this.createUser.bind(this);
-    this.updateUser = this.updateUser.bind(this);
-    this.deleteUser = this.deleteUser.bind(this);
-    this.addFile = this.addFile.bind(this);
-    this.removeFile = this.removeFile.bind(this);
-  }
+  ) {}
 
   @ControllerMethod()
   async getAllUsers(req: Request, res: Response): Promise<void> {
