+++ server/tests/utils/authentication.ts
@@ -1,0 +1,3 @@
+function mockAuthentication() {
+  
+}
